+++ .github/workflows/test-rekor-integration.yml
@@ -2,13 +2,7 @@ name: Test Rekor Integration
 
 on:
   push:
-    branches:
-      - main
-    paths:
-      - 'pkg/test/rekor_integration/**'
   pull_request:
-    paths:
-      - 'pkg/test/rekor_integration/**'
 
 jobs:
   rekor_integration_tests:
@@ -24,5 +18,4 @@ jobs:
           go-version: 1.24
 
       - name: Run Rekor Integration tests
-        working-directory: ./pkg/test/rekor_integration
-        run: go test -v -timeout=1m
+        run: ./test/integration/rekor_monitor_integration_test.sh
