+++ test/integration/rekor_monitor_integration_test.go
@@ -1,0 +1,218 @@
+package e2e
+
+import (
+	"context"
+	"fmt"
+	"io"
+	"net"
+	"net/http"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"strings"
+	"syscall"
+	"testing"
+	"time"
+)
+
+type MonitorExpectations struct {
+	ExpectErrorLog       bool
+	ExpectedFailureCount int
+	ExpectedTotalCount   int
+}
+
+func TestMonitorWithValidCheckpoint(t *testing.T) {
+	ctx, binary, checkpointFile, monitorPort := setupTest(t)
+
+	t.Run("start_monitor_and_check_initial_metrics", func(t *testing.T) {
+		runMonitorAndValidate(t, ctx, binary, checkpointFile, monitorPort, MonitorExpectations{
+			ExpectErrorLog:       false,
+			ExpectedFailureCount: 0,
+			ExpectedTotalCount:   1,
+		})
+	})
+}
+
+func TestTamperedCheckpoint(t *testing.T) {
+	ctx, binary, checkpointFile, monitorPort := setupTest(t)
+
+	t.Run("validate_and_tamper_checkpoint_file", func(t *testing.T) {
+		content, err := os.ReadFile(checkpointFile)
+		if err != nil {
+			t.Fatalf("failed to read checkpoint file: %v", err)
+		}
+		contentStr := string(content)
+		wasNewlineTrailing := strings.HasSuffix(contentStr, "\n")
+		contentStr = strings.TrimSuffix(contentStr, "\n")
+		wasEscapedTrailing := strings.HasSuffix(contentStr, "\\n")
+		contentStr = strings.TrimSuffix(contentStr, "\\n")
+		lines := strings.SplitN(contentStr, "\\n", 4)
+		if len(lines) < 3 {
+			t.Fatalf("invalid checkpoint file format: expected at least 3 lines, got %d", len(lines))
+		}
+		lines[2] = lines[2] + "tampered"
+		tamperedContent := strings.Join(lines, "\\n")
+		if wasEscapedTrailing {
+			tamperedContent += "\\n"
+		}
+		tamperedBytes := []byte(tamperedContent)
+		if wasNewlineTrailing {
+			tamperedBytes = append(tamperedBytes, '\n')
+		}
+		if err := os.WriteFile(checkpointFile, tamperedBytes, 0644); err != nil {
+			t.Fatalf("failed to write tampered checkpoint: %v", err)
+		}
+	})
+
+	runMonitorAndValidate(t, ctx, binary, checkpointFile, monitorPort, MonitorExpectations{
+		ExpectErrorLog:       true,
+		ExpectedFailureCount: 1,
+		ExpectedTotalCount:   1,
+	})
+}
+
+// setupTest prepares the test environment: builds the binary, initializes the checkpoint file,
+// and allocates a port. Returns the context, binary path, checkpoint file path, and monitor port.
+func setupTest(t *testing.T) (context.Context, string, string, string) {
+	t.Helper()
+
+	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
+	t.Cleanup(cancel)
+
+	buildDir := t.TempDir()
+	binary := filepath.Join(buildDir, "rekor_monitor")
+	buildCmd := exec.Command("go", "build", "-o", binary, "../../cmd/rekor_monitor/main.go")
+	buildCmd.Stdout = os.Stdout
+	buildCmd.Stderr = os.Stderr
+	if err := buildCmd.Run(); err != nil {
+		t.Fatalf("failed to build binary: %v", err)
+	}
+
+	// Initialize checkpoint file
+	dataDir := t.TempDir()
+	checkpointFile := filepath.Join(dataDir, "checkpoint_log.txt")
+
+	t.Run("generate initial checkpoint file", func(t *testing.T) {
+		initCmd := exec.CommandContext(ctx, binary,
+			"--once",
+			"--file", checkpointFile,
+			"--url", "http://localhost:3000",
+		)
+		initCmd.Stdout = os.Stdout
+		initCmd.Stderr = os.Stderr
+		err := initCmd.Run()
+		if err == nil {
+			t.Fatal("expected error on initial run due to no start index")
+		}
+
+		if _, err := os.Stat(checkpointFile); err != nil {
+			t.Fatalf("checkpoint file not created: %v", err)
+		}
+	})
+
+	// Allocate port
+	listener, err := net.Listen("tcp", "localhost:0")
+	if err != nil {
+		t.Fatalf("failed to find free port: %v", err)
+	}
+	monitorPort := fmt.Sprintf("%d", listener.Addr().(*net.TCPAddr).Port)
+	listener.Close()
+
+	return ctx, binary, checkpointFile, monitorPort
+}
+
+// runMonitorAndValidate runs the monitor, captures logs, fetches metrics, and validates expected errors and metrics.
+func runMonitorAndValidate(t *testing.T, ctx context.Context, binary, checkpointFile, monitorPort string, exp MonitorExpectations) {
+	t.Helper()
+
+	t.Run("start_monitor", func(t *testing.T) {
+		runCmd := exec.CommandContext(ctx, binary,
+			"--once=false",
+			"--interval=2s",
+			"--file", checkpointFile,
+			"--url", "http://localhost:3000",
+			"--monitor-port", monitorPort,
+		)
+
+		stdoutPipe, err := runCmd.StdoutPipe()
+		if err != nil {
+			t.Fatalf("failed to get stdout pipe: %v", err)
+		}
+		stderrPipe, err := runCmd.StderrPipe()
+		if err != nil {
+			t.Fatalf("failed to get stderr pipe: %v", err)
+		}
+		if err := runCmd.Start(); err != nil {
+			t.Fatalf("failed to start monitor: %v", err)
+		}
+		defer runCmd.Process.Kill()
+
+		var logs strings.Builder
+		go io.Copy(&logs, stdoutPipe)
+		go io.Copy(&logs, stderrPipe)
+
+		time.Sleep(1500 * time.Millisecond)
+
+		var metricsStr string
+		t.Run("fetch_metrics", func(t *testing.T) {
+			for i := 0; i < 5; i++ {
+				resp, err := http.Get(fmt.Sprintf("http://localhost:%s/metrics", monitorPort))
+				if err == nil {
+					body, err := io.ReadAll(resp.Body)
+					resp.Body.Close()
+					if err == nil {
+						metricsStr = string(body)
+						break
+					}
+					t.Logf("failed to read metrics body: %v", err)
+				}
+				time.Sleep(500 * time.Millisecond)
+			}
+			if metricsStr == "" {
+				t.Fatalf("failed to query metrics after retries")
+			}
+		})
+
+		t.Run("validate_logs_and_metrics", func(t *testing.T) {
+			if exp.ExpectErrorLog {
+				if !strings.Contains(logs.String(), "error running consistency check") {
+					t.Errorf("expected error log not found:\n%s", logs.String())
+				}
+			} else {
+				if strings.Contains(logs.String(), "error") {
+					t.Errorf("unexpected error found in logs:\n%s", logs.String())
+				}
+			}
+
+			failMetric := fmt.Sprintf("log_index_verification_failure %d", exp.ExpectedFailureCount)
+			if !strings.Contains(metricsStr, failMetric) {
+				t.Errorf("expected failure metric '%s' not found:\n%s", failMetric, metricsStr)
+			}
+
+			totalMetric := fmt.Sprintf("log_index_verification_total %d", exp.ExpectedTotalCount)
+			if !strings.Contains(metricsStr, totalMetric) {
+				t.Errorf("expected total metric '%s' not found:\n%s", totalMetric, metricsStr)
+			}
+		})
+
+		t.Run("stop_monitor", func(t *testing.T) {
+			if err := runCmd.Process.Signal(syscall.SIGTERM); err != nil {
+				t.Fatalf("failed to send SIGTERM to monitor: %v", err)
+			}
+			select {
+			case <-time.After(5 * time.Second):
+				t.Fatalf("monitor did not exit within 5 seconds after SIGTERM")
+			case err := <-func() chan error {
+				errChan := make(chan error)
+				go func() {
+					errChan <- runCmd.Wait()
+				}()
+				return errChan
+			}():
+				if err != nil && !strings.Contains(err.Error(), "signal: terminated") {
+					t.Fatalf("monitor did not exit cleanly: %v", err)
+				}
+			}
+		})
+	})
+}
